Refactor the ParcelTrackerResultsHandler
This change is required after the recent refactor of
ParcelRepository::findTrackingDetailsByTrackingNumber. Well, strictly
speaking the refactor in ParcelTrackerResultsHandler isn't but the
refactor of its template is. Specifically, the template has been updated
to use the Parcel entity, along with its related entities properly, so
that the relevant/required information is displayed, if available.

The refactor in ParcelTrackerResultsHandler is mainly focused on
cleaning it up and making it a bit easier to read.