Revise the list of environment variables used in the application
This change adds three more (SENDGRID_SENDER_EMAIL, SENDGRID_SENDER_NAME,
and TWILIO_PHONE_NUMBER). It adds them to .env.example so that it's
clear that they're required and what they're for. It defines them in
compose.yml to assist with loading them on container boot, and it uses
them in the SendGrid and Twilio autoload config files.