Reorganise and improve the listeners
This change both reorganises the directory structure of the listeners,
it renames them so that they're as meaningfully named as possible. It's
my intention that, with this commit, the code is as easy to intuit and
understand as possible. The code also was helped by having PHP CS
Fixer and PHPStan review the code and have changes made in response to
that review.